requirements.txt
asgiref
awesomeversion
beautifulsoup4
cryptography
cloudinary
dj-database-url
Django==3.1.7
django-cors-headers
django-extensions
django-js-asset
gunicorn
psycopg2-binary
pytz
PyYAML
requests
six
whitenoise
django-rest-auth[with_social]
djangorestframework
markdown
django-filter
django-datatables-view
Pillow
pyimgur
